refactor: address review on canonical phase guards, structure, and docs
Extract the stochastic-op scan into HirModule::is_deterministic, promote
the canonical-phase consistency asserts to unconditional throws so an
unhandled gate kind or corrupted log fails loudly instead of poisoning
global_weight with NaN in release builds, and move the canonical-phase
oracle tests into their own test_canonical_phase.cc mirroring the new
util TU (dense helpers consolidated in test_helpers.h).

Document the statevector phase contract: exact including global phase
for measurement/noise-free programs; after measurements or noise, exact
up to a global phase, which a collapsed state does not physically carry.
