+++ tests/test_backend.cc
@@ -1558,108 +1558,3 @@ TEST_CASE("Lower: queued virtual gates affect later noise masks") {
     CHECK(cv.z() == Z(0));
     CHECK_THAT(ch.prob, Catch::Matchers::WithinAbs(0.25, 1e-12));
 }
-
-// =============================================================================
-// Frame composition canonical phase -- dense stim oracle
-// =============================================================================
-
-namespace {
-
-using clifft::test::dense_matmul;
-using clifft::test::dense_tableau_matrix;
-using clifft::test::DenseMatrix;
-
-// Canonical matrix of a single pending frame gate, little-endian basis.
-DenseMatrix dense_pending_gate(const PendingGate& g, size_t n) {
-    const uint64_t dim = uint64_t{1} << n;
-    constexpr std::complex<double> kI{0.0, 1.0};
-    const double inv_sqrt2 = 1.0 / std::sqrt(2.0);
-
-    DenseMatrix m(dim * dim, {0.0, 0.0});
-    for (uint64_t c = 0; c < dim; ++c) {
-        const bool b1 = ((c >> g.axis_1) & 1) != 0;
-        const bool b2 = ((c >> g.axis_2) & 1) != 0;
-        switch (g.type) {
-            case PendingGateType::S:
-                m[c * dim + c] = b1 ? kI : 1.0;
-                break;
-            case PendingGateType::CZ:
-                m[c * dim + c] = (b1 && b2) ? -1.0 : 1.0;
-                break;
-            case PendingGateType::CNOT:
-                m[(b1 ? c ^ (uint64_t{1} << g.axis_2) : c) * dim + c] = 1.0;
-                break;
-            case PendingGateType::SWAP: {
-                uint64_t r = c;
-                if (b1 != b2) {
-                    r ^= (uint64_t{1} << g.axis_1) | (uint64_t{1} << g.axis_2);
-                }
-                m[r * dim + c] = 1.0;
-                break;
-            }
-            case PendingGateType::H: {
-                const uint64_t r0 = c & ~(uint64_t{1} << g.axis_1);
-                const uint64_t r1 = c | (uint64_t{1} << g.axis_1);
-                m[r0 * dim + c] = inv_sqrt2;
-                m[r1 * dim + c] = b1 ? -inv_sqrt2 : inv_sqrt2;
-                break;
-            }
-        }
-    }
-    return m;
-}
-
-}  // namespace
-
-TEST_CASE("Backend: frame composition phase matches dense canonical oracle") {
-    // For target tableau T and frame gates u_1..u_J (application order),
-    // composed = tableau(T * (u_J...u_1)^{-1}) is what lower() stores; the
-    // chained phase must satisfy
-    //   canonical(composed) * u_J * ... * u_1 == phase * canonical(T)
-    // componentwise against stim's dense matrices.
-    std::mt19937_64 rng(0x141);
-
-    for (int trial = 0; trial < 128; ++trial) {
-        CAPTURE(trial);
-        const size_t n = 2 + static_cast<size_t>(trial % 2);
-        const auto target = stim::Tableau<kStimWidth>::random(n, rng);
-
-        std::vector<PendingGate> log;
-        const size_t num_gates = 1 + rng() % 8;
-        for (size_t i = 0; i < num_gates; ++i) {
-            PendingGate g{};
-            g.type = static_cast<PendingGateType>(rng() % 5);
-            g.axis_1 = static_cast<uint16_t>(rng() % n);
-            g.axis_2 = static_cast<uint16_t>(rng() % n);
-            if ((g.type == PendingGateType::CNOT || g.type == PendingGateType::CZ ||
-                 g.type == PendingGateType::SWAP) &&
-                g.axis_2 == g.axis_1) {
-                g.axis_2 = static_cast<uint16_t>((g.axis_1 + 1) % n);
-            }
-            log.push_back(g);
-        }
-
-        // Reproduce lower()'s composition through the production frame.
-        VirtualFrame frame(static_cast<uint32_t>(n), 4);
-        for (const auto& g : log) {
-            frame.append_gate(g);
-        }
-        const stim::Tableau<kStimWidth> v_cum_inv = frame.mutable_materialized_tableau().inverse();
-        const stim::Tableau<kStimWidth> composed = v_cum_inv.then(target);
-
-        const auto phase = frame_composition_phase(composed, log, target);
-
-        const uint64_t dim = uint64_t{1} << n;
-        DenseMatrix lhs = dense_tableau_matrix(composed);
-        for (auto it = log.rbegin(); it != log.rend(); ++it) {
-            lhs = dense_matmul(lhs, dense_pending_gate(*it, n), dim);
-        }
-        const DenseMatrix rhs = dense_tableau_matrix(target);
-        for (uint64_t i = 0; i < dim * dim; ++i) {
-            CAPTURE(i);
-            const auto expected = phase * rhs[i];
-            REQUIRE_THAT(lhs[i].real(), Catch::Matchers::WithinAbs(expected.real(), 1e-5));
-            REQUIRE_THAT(lhs[i].imag(), Catch::Matchers::WithinAbs(expected.imag(), 1e-5));
-        }
-    }
-}
