+++ tests/test_optimizer.cc
@@ -720,30 +720,10 @@ TEST_CASE("Peephole: commuting NOISE does not bypass EXP_VAL barrier", "[optimiz
 
 namespace {
 
+using clifft::test::dense_axis_rotation;
 using clifft::test::dense_matmul;
 using clifft::test::dense_tableau_matrix;
 using clifft::test::DenseMatrix;
-
-// Dense matrix of the projector-form rotation Pi_+ + e^{i*alpha*pi} Pi_- on
-// the signed Pauli (x, z, sign) over n qubits, little-endian basis order.
-// The fused S/S_dag the peephole absorbs is alpha = 0.5 / 1.5.
-DenseMatrix dense_axis_rotation(uint64_t x, uint64_t z, bool sign, double alpha, size_t n) {
-    const uint64_t dim = uint64_t{1} << n;
-    constexpr std::complex<double> kIPow[4] = {{1, 0}, {0, 1}, {-1, 0}, {0, -1}};
-    const std::complex<double> eig{std::cos(alpha * std::numbers::pi),
-                                   std::sin(alpha * std::numbers::pi)};
-    const std::complex<double> a = (1.0 + eig) / 2.0;
-    const std::complex<double> b = (1.0 - eig) / 2.0;
-
-    DenseMatrix r(dim * dim, {0.0, 0.0});
-    for (uint64_t c = 0; c < dim; ++c) {
-        r[c * dim + c] += a;
-        uint32_t phase_idx = (sign ? 2U : 0U) + static_cast<uint32_t>(std::popcount(x & z)) +
-                             2U * (static_cast<uint32_t>(std::popcount(c & z)) & 1U);
-        r[(c ^ x) * dim + c] += b * kIPow[phase_idx & 3U];
-    }
-    return r;
-}
 
 // Dense value of an HIR module: global_weight * canonical(final_tableau)
 // applied after the op stream. T_GATE is projector-form on the signed axis;
@@ -774,51 +754,6 @@ DenseMatrix dense_hir_value(const HirModule& hir) {
 }
 
 }  // namespace
-
-TEST_CASE("Peephole: S absorption phase matches dense canonical oracle", "[optimizer]") {
-    // Fuzz s_absorption_phase against stim's to_flat_unitary_matrix: for a
-    // random frame U_C and random signed Pauli P, the exact rotation R must
-    // satisfy canonical(U_C) * R == phase * canonical(U_C * S_P)
-    // componentwise, where phase is the value folded into global_weight.
-    std::mt19937_64 rng(0x139);
-
-    for (int trial = 0; trial < 256; ++trial) {
-        CAPTURE(trial);
-        const size_t n = 1 + static_cast<size_t>(trial % 4);
-        const uint64_t qubit_mask = (uint64_t{1} << n) - 1;
-        auto tab = stim::Tableau<kStimWidth>::random(n, rng);
-
-        uint64_t x = 0;
-        uint64_t z = 0;
-        while (x == 0 && z == 0) {
-            x = rng() & qubit_mask;
-            z = rng() & qubit_mask;
-        }
-        const bool sign = (rng() & 1) != 0;
-        const bool is_dagger = (rng() & 1) != 0;
-
-        const uint64_t x_words[1] = {x};
-        const uint64_t z_words[1] = {z};
-        MaskView x_v{std::span<const uint64_t>(x_words)};
-        MaskView z_v{std::span<const uint64_t>(z_words)};
-
-        auto updated = tab;
-        internal::apply_s_to_tableau(updated, x_v, z_v, sign, is_dagger);
-        const auto phase = internal::s_absorption_phase(tab, updated, x_v, z_v, sign, is_dagger);
-
-        const uint64_t dim = uint64_t{1} << n;
-        const auto lhs =
-            dense_matmul(dense_tableau_matrix(tab),
-                         dense_axis_rotation(x, z, sign, is_dagger ? 1.5 : 0.5, n), dim);
-        const auto m_new = dense_tableau_matrix(updated);
-        for (uint64_t i = 0; i < dim * dim; ++i) {
-            const auto rhs = phase * m_new[i];
-            CAPTURE(i);
-            REQUIRE_THAT(lhs[i].real(), Catch::Matchers::WithinAbs(rhs.real(), 1e-5));
-            REQUIRE_THAT(lhs[i].imag(), Catch::Matchers::WithinAbs(rhs.imag(), 1e-5));
-        }
-    }
-}
 
 TEST_CASE("Peephole: pass preserves dense HIR value on random circuits", "[optimizer]") {
     // Value-preservation fuzz: the pass must keep
@@ -891,82 +826,4 @@ TEST_CASE("Peephole: S absorption on wide multi-word Pauli axes", "[optimizer]")
 
     REQUIRE(pass.fusions() == 2);
     REQUIRE(hir.ops.empty());
-}
-
-TEST_CASE("Peephole: S absorption phase is invariant under qubit embedding", "[optimizer]") {
-    // Multi-word phase oracle. The canonical phase delta factorizes over
-    // tensor products with identity: the Choi stabilizer group, its
-    // minimum-integer support index, and the amplitude walk all decompose
-    // over disjoint qubit sets, so embedding the frame and the Pauli at a
-    // qubit offset must leave s_absorption_phase unchanged. Offsets
-    // straddling bit 64 of a 72-qubit frame drive the 144-bit Choi indices
-    // across word boundaries, pinning the multi-word arithmetic to the
-    // dense-validated small-frame value.
-    std::mt19937_64 rng(0xC0FFEE);
-    constexpr size_t k = 3;
-    constexpr size_t n_wide = 72;
-
-    for (int trial = 0; trial < 64; ++trial) {
-        CAPTURE(trial);
-        const size_t offset = 60 + static_cast<size_t>(trial % 8);
-        CAPTURE(offset);
-        auto small_tab = stim::Tableau<kStimWidth>::random(k, rng);
-
-        uint64_t x = 0;
-        uint64_t z = 0;
-        while (x == 0 && z == 0) {
-            x = rng() & 7;
-            z = rng() & 7;
-        }
-        const bool sign = (rng() & 1) != 0;
-        const bool is_dagger = (rng() & 1) != 0;
-
-        auto wide_tab = stim::Tableau<kStimWidth>::identity(n_wide);
-        for (size_t a = 0; a < k; ++a) {
-            auto wx = wide_tab.xs[offset + a];
-            auto wz = wide_tab.zs[offset + a];
-            wx.xs[offset + a] = false;
-            wz.zs[offset + a] = false;
-            for (size_t b = 0; b < k; ++b) {
-                wx.xs[offset + b] = small_tab.xs[a].xs[b];
-                wx.zs[offset + b] = small_tab.xs[a].zs[b];
-                wz.xs[offset + b] = small_tab.zs[a].xs[b];
-                wz.zs[offset + b] = small_tab.zs[a].zs[b];
-            }
-            wx.sign = static_cast<bool>(small_tab.xs[a].sign);
-            wz.sign = static_cast<bool>(small_tab.zs[a].sign);
-        }
-
-        const uint64_t x_small_words[1] = {x};
-        const uint64_t z_small_words[1] = {z};
-        MaskView x_small{std::span<const uint64_t>(x_small_words)};
-        MaskView z_small{std::span<const uint64_t>(z_small_words)};
-
-        uint64_t x_wide_words[2] = {0, 0};
-        uint64_t z_wide_words[2] = {0, 0};
-        for (size_t b = 0; b < k; ++b) {
-            const size_t bit = offset + b;
-            if ((x >> b) & 1) {
-                x_wide_words[bit / 64] |= uint64_t{1} << (bit % 64);
-            }
-            if ((z >> b) & 1) {
-                z_wide_words[bit / 64] |= uint64_t{1} << (bit % 64);
-            }
-        }
-        MaskView x_wide{std::span<const uint64_t>(x_wide_words)};
-        MaskView z_wide{std::span<const uint64_t>(z_wide_words)};
-
-        auto small_updated = small_tab;
-        internal::apply_s_to_tableau(small_updated, x_small, z_small, sign, is_dagger);
-        const auto phase_small = internal::s_absorption_phase(small_tab, small_updated, x_small,
-                                                              z_small, sign, is_dagger);
-
-        auto wide_updated = wide_tab;
-        internal::apply_s_to_tableau(wide_updated, x_wide, z_wide, sign, is_dagger);
-        const auto phase_wide =
-            internal::s_absorption_phase(wide_tab, wide_updated, x_wide, z_wide, sign, is_dagger);
-
-        REQUIRE_THAT(phase_wide.real(), Catch::Matchers::WithinAbs(phase_small.real(), 1e-9));
-        REQUIRE_THAT(phase_wide.imag(), Catch::Matchers::WithinAbs(phase_small.imag(), 1e-9));
-    }
 }
